Move crypto details screen to presentation package
