+++ app/src/main/java/com/example/traders/presentation/cryptoDetailsScreen/CryptoItemViewModel.kt
@@ -1,10 +1,9 @@
-package com.example.traders.singleCryptoScreen
+package com.example.traders.presentation.cryptoDetailsScreen
 
 import androidx.lifecycle.SavedStateHandle
 import com.example.traders.BaseViewModel
 import com.example.traders.database.FavouriteCrypto
 import com.example.traders.network.repository.CryptoRepository
-import com.example.traders.singleCryptoScreen.priceStatisticsTab.Hilt_CryptoPriceStatisticsFragment
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.MutableSharedFlow
