+++ app/src/main/java/com/example/traders/presentation/cryptoDetailsScreen/chartTab/CandleType.kt
@@ -1,4 +1,4 @@
-package com.example.traders.singleCryptoScreen.chartTab
+package com.example.traders.presentation.cryptoDetailsScreen.chartTab
 
 enum class CandleType(val numDays: Long, val candleInterval: String) {
     DAILY(90, "1d"),
