+++ app/src/main/java/com/example/traders/presentation/cryptoDetailsScreen/descriptionTab/DescriptionState.kt
@@ -1,4 +1,4 @@
-package com.example.traders.singleCryptoScreen.descriptionTab
+package com.example.traders.presentation.cryptoDetailsScreen.descriptionTab
 
 data class DescriptionState(
     val projectInfoDesc: String? = null,
