+++ app/src/main/java/com/example/traders/presentation/dialogs/buyDialog/BuyState.kt
@@ -1,4 +1,4 @@
-package com.example.traders.dialogs.buyDialog
+package com.example.traders.presentation.dialogs.buyDialog
 
 import com.example.traders.database.Crypto
 import com.example.traders.dialogs.DialogValidationMessage
