+++ app/src/main/res/navigation/watchlist.xml
@@ -22,7 +22,7 @@
 
     <fragment
         android:id="@+id/cryptoItemFragment"
-        android:name="com.example.traders.singleCryptoScreen.CryptoItemFragment"
+        android:name="com.example.traders.presentation.cryptoDetailsScreen.CryptoItemFragment"
         android:label="CryptoItem" >
         <argument
             android:name="slug"
