+++ app/src/test/java/com/example/traders/dialogs/buyDialog/BuyDialogViewModelTest.kt
@@ -8,6 +8,8 @@ import com.example.traders.network.repository.CryptoRepository
 import com.example.traders.utils.roundNum
 import com.example.traders.TestDispatcherRule
 import com.example.traders.database.FixedCryptoList
+import com.example.traders.presentation.dialogs.buyDialog.BuyDialogEvent
+import com.example.traders.presentation.dialogs.buyDialog.BuyDialogViewModel
 import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.whenever
 import kotlinx.coroutines.ExperimentalCoroutinesApi
