+++ app/src/test/java/com/example/traders/singleCryptoScreen/descriptionTab/CryptoDescriptionViewModelTest.kt
@@ -6,6 +6,7 @@ import com.example.traders.network.repository.CryptoRepository
 import com.example.traders.TestDispatcherRule
 import com.example.traders.database.FixedCryptoList
 import com.example.traders.network.models.cryptoDescData.CryptoDescData
+import com.example.traders.presentation.cryptoDetailsScreen.descriptionTab.CryptoDescriptionViewModel
 import com.nhaarman.mockitokotlin2.any
 import com.nhaarman.mockitokotlin2.mock
 import com.nhaarman.mockitokotlin2.whenever
